initial layout and folder structure
