+++ electron/preload.ts
@@ -38,8 +38,3 @@ const api = {
   }
 };
 contextBridge.exposeInMainWorld('Main', api);
-/**
- * Using the ipcRenderer directly in the browser through the contextBridge ist not really secure.
- * I advise using the Main/api way !!
- */
-contextBridge.exposeInMainWorld('ipcRenderer', ipcRenderer);
